offline payment done successfully

+++ database/migrations/2024_02_25_174402_create_orders_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('user_id')->nullable();
             $table->string('code',250)->unique()->nullable();
-            $table->double('amount')->nullable();
+            $table->unsignedBigInteger('amount')->nullable();
             $table->timestamps();
         });
         \Illuminate\Support\Facades\DB::update('alter table orders AUTO_INCREMENT = 100000');
